submitting fixes
