+++ src/app/student/specific/pages/specific-news/specific-news.component.html
@@ -41,7 +41,7 @@
                   Links here:
                   <ul>
                     <li *ngFor="let links of this.pageService.hyperLinks">
-                      <a href="{{links}}">{{links}}</a>
+                      <a href="{{links}}" target="_blank">{{links}}</a>
                     </li>
                   </ul>
                 </div>
@@ -72,7 +72,7 @@
                           Links here:
                           <ul>
                             <li *ngFor="let links of this.pageService.hyperLinks">
-                              <a href="{{links}}">{{links}}</a>
+                              <a href="{{links}}" target="_blank">{{links}}</a>
                             </li>
                           </ul>
                         </div>
