+++ src/app/student/specific/pages/specific-quest-map/subpages/map-chart/map-chart.component.html
@@ -122,7 +122,7 @@
 					<h5>Links:</h5>
 					<ol class="list">
 						<li *ngIf="!this.pageService.getPossibleLinks(this.questClicked.getQuestDescription())" style="color: red"><i>None</i></li>
-						<li *ngFor="let links of this.pageService.hyperLinks"><a href="{{links}}">{{links}}</a></li>
+						<li *ngFor="let links of this.pageService.hyperLinks"><a href="{{links}}" target="_blank">{{links}}</a></li>
 					</ol>
 				</td>
 			</tr>
