+++ src/app/student/specific/pages/specific-quest-map/subpages/map-chart/map-chart.component.ts
@@ -294,13 +294,16 @@ export class MapChartComponent implements OnInit {
 		}
 
 		var HTMLchart = document.getElementById("quest-map");
-		var ctx = (<HTMLCanvasElement>HTMLchart).getContext("2d");
 
-		let cc: any = {
-			data: QM,
-			options: options
-		};
-		this.chart = new Chart(ctx, cc);
+		if(<HTMLCanvasElement>HTMLchart) {
+			var ctx = (<HTMLCanvasElement>HTMLchart).getContext("2d");
+
+			let cc: any = {
+				data: QM,
+				options: options
+			};
+			this.chart = new Chart(ctx, cc);
+		}
 
 		//this.onChartClick(HTMLchart, chart, this.chartWidth, this.chartHeight, xTick, yTick);
