+++ src/app/teacher/specific/pages/grades/copy/grades.component.css
@@ -1,0 +1,4 @@
+/*Table styles*/
+#hor-zebra {
+    width: 70%;
+}
