+++ src/app/teacher/specific/pages/grades/grades-navbar/grades-navbar.component.css
@@ -1,0 +1,31 @@
+.navbarLine {
+    content: ' ';
+    margin-top: -5px;
+    display: block;
+    width: 95%;
+    border: 2px solid #d0d0d0;
+    border-radius: 4px;
+    -webkit-border-radius: 4px;
+    -moz-border-radius: 4px;
+
+    box-shadow: inset 0 1px 1px rgba(0, 0, 0, .05);
+    -webkit-box-shadow: inset 0 1px 1px rgba(0, 0, 0, .05);
+    -moz-box-shadow: inset 0 1px 1px rgba(0, 0, 0, .05);
+}
+
+.qMapLink {
+    text-decoration: none;
+    color: black;
+    font-size: 30px;
+}
+
+.navbar {
+    width: 95%;
+    padding-right: 10%;
+    padding-left: 10%;
+}
+
+.verticalBar {
+    border-left: 2px solid black;
+    height: 30px;
+}
