+++ src/app/teacher/specific/pages/grades/grades-navbar/grades-navbar.component.html
@@ -1,0 +1,9 @@
+<div class="navbar">
+	<a class="qMapLink" [routerLink]="['./']">Grades</a>
+    <div class="verticalBar"></div>
+    <a class="qMapLink" [routerLink]="['./subpages/graph-summary']">Graph</a>
+    <div class="verticalBar"></div>
+	<a class="qMapLink" [routerLink]="['./subpages/submitted-quests']">Submitted Quests</a>
+</div>
+
+<span class="navbarLine"></span>
