+++ src/app/teacher/specific/pages/grades/grades-navbar/grades-navbar.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-grades-navbar',
+  templateUrl: './grades-navbar.component.html',
+  styleUrls: ['./grades-navbar.component.css']
+})
+export class GradesNavbarComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
