+++ src/app/teacher/specific/pages/grades/grades.component.css
@@ -1,4 +1,3 @@
-/*Table styles*/
-#hor-zebra {
-    width: 70%;
+.side-body {
+    padding-top: 25px;
 }
