+++ src/app/teacher/specific/pages/grades/index.ts
@@ -1,0 +1,7 @@
+export * from './grades-navbar/grades-navbar.component';
+export * from './subpages/submitted-quests/submitted-quests.component';
+export * from './subpages/grades-summary/grades-summary.component';
+export * from './subpages/graph-summary/graph-summary.component';
+export {
+    StudentGrades
+} from './subpages/grades-summary/grades-summary.component';
