+++ src/app/teacher/specific/pages/grades/subpages/grades-summary/grades-summary.component.css
@@ -1,0 +1,4 @@
+/*Table styles*/
+#hor-zebra {
+    width: 70%;
+}
