+++ src/app/teacher/specific/pages/specific-news/specific-news.component.html
@@ -52,7 +52,7 @@
 							Links here:
 							<ul>
 								<li *ngFor="let links of this.pageService.hyperLinks">
-									<a href="{{links}}">{{links}}</a>
+									<a href="{{links}}" target="_blank">{{links}}</a>
 								</li>
 							</ul>
 						</div>
@@ -78,7 +78,7 @@
 											Links here:
 											<ul>
 												<li *ngFor="let links of this.pageService.hyperLinks">
-													<a href="{{links}}">{{links}}</a>
+													<a href="{{links}}" target="_blank">{{links}}</a>
 												</li>
 											</ul>
 										</div>
