+++ src/app/teacher/specific/pages/specific-quest-map/specific-quest-map-navbar/specific-quest-map-navbar.component.ts
@@ -1,13 +1,7 @@
 import {
 	Component,
-	OnInit,
-	Input
+	OnInit
 } from '@angular/core';
-
-import { SectionService } from 'shared/services/section.service';
-
-import { ActivatedRoute } from '@angular/router';
-import { Section } from 'shared/models/section';
 
 @Component({
 	selector: 'app-specific-quest-map-navbar',
