+++ src/app/teacher/specific/pages/specific-quest-map/subpages/map-chart/map-chart.component.html
@@ -67,7 +67,7 @@
 					<h5>Links:</h5>
 					<ol>
 						<li *ngIf="!this.pageService.getPossibleLinks(this.questClicked.getQuestDescription())" style="color: red"><i>None</i></li>
-						<li *ngFor="let links of this.pageService.hyperLinks"><a href="{{links}}">{{links}}</a></li>
+						<li *ngFor="let links of this.pageService.hyperLinks"><a href="{{links}}" target="_blank">{{links}}</a></li>
 					</ol>
 				</td>
 			</tr>
